float removed, char split

+++ c_source/vga.h
@@ -1,7 +1,7 @@
 #ifndef __VGA__
 #define __VGA__
 
-void printNumber(float number, int posx, int posy);
+void printNumber(int number, int posx, int posy);
 void printOperand(char c);
 void printError(void);
 void clearDisplay(void);
